+++ AzureCommunicationUI/AzureCommunicationUIDemoApp/Tests/Utilities/XCUIElementExtension.swift
@@ -32,13 +32,13 @@ extension XCUIElement {
 }
 
 extension XCTestCase {
-  func wait(for element: XCUIElement, timeout: TimeInterval = 10) {
+    func wait(for element: XCUIElement, timeout: TimeInterval = 10.0) {
     let predicate = NSPredicate(format: "exists == true")
     let expectation = expectation(for: predicate, evaluatedWith: element, handler: nil)
     wait(for: [expectation], timeout: timeout)
   }
 
-    func waitEnabled(for element: XCUIElement, timeout: TimeInterval = 10) {
+    func waitEnabled(for element: XCUIElement, timeout: TimeInterval = 10.0) {
       let predicate = NSPredicate(format: "enabled == true")
       let expectation = expectation(for: predicate, evaluatedWith: element, handler: nil)
       wait(for: [expectation], timeout: timeout)
